Adding a custom moving 5 median to find the dt for calculations
3 effects were observed with C2:
(1) When tabbing/stopping&starting the scene 1-2 'fast' frames appear
    (dt < 5ms) which - since detected movement may be based on previous
    frame length - leads to spikes is calculated external speed
(2) When switching windows to/from browser 1 'timeout' frame appears
    (dt = 100ms) which has the same speed-calculation problem
Both of these can be mitigated by a moving median.
It is ideal because it tolerates upto 2 extreme values. A 3 median was not enough.
(3) At the very start of a scene 1-2 frames might drop in with dt = 0.
    These are no concern atm.

The median is also the most computationally expensive measure i tried so far
i guess, since it involves sorting a 5-element list. Still shouldn't be too bad, yet.